LayoutTests/js/dom/unhandled-promise-rejection-basic-expected.txt

Test unhandled promise rejection event.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS error.type is "unhandledrejection"
PASS error.cancelable is true
PASS error.promise is promise
PASS error.reason is "ERROR"
PASS successfullyParsed is true

TEST COMPLETE
